[minor fixes and changes]
	- align the return type of the read(), write() declaration and
	  definition
	- update the type of some variables
	- add missing initialisations of static variables
	- disable optimisation for host targets in debug builds
	- set errno if kmalloc runs out of memory
	- add file and line number to test result log entries
	- indicate the overall test result through the application return
	  value

+++ kernel/fs/syscall.c
@@ -163,7 +163,7 @@ static int sc_hdlr_close(void *_p){
 }
 
 static int sc_hdlr_read(void *_p){
-	int r;
+	ssize_t r;
 	char buf[((sc_fs_t*)(_p))->data_len];
 	sc_fs_t *p;
 	fs_filed_t *fd;
